+++ packages/assets-controllers/src/token-prices-service/index.test.ts
@@ -5,6 +5,7 @@ describe('token-prices-service', () => {
     expect(Object.keys(allExports)).toMatchInlineSnapshot(`
       Array [
         "CodefiTokenPricesServiceV2",
+        "SUPPORTED_CHAIN_IDS",
       ]
     `);
   });