+++ packages/controller-utils/src/constants.ts
@@ -108,6 +108,8 @@ export const BUILT_IN_NETWORKS = {
 export const OPENSEA_PROXY_URL =
   'https://proxy.metafi.codefi.network/opensea/v1/api/v2';
 
+export const NFT_API_BASE_URL = 'https://nft.api.cx.metamask.io';
+
 // Default origin for controllers
 export const ORIGIN_METAMASK = 'metamask';