+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -54,6 +54,7 @@ import type {
   SimulationData,
 } from './types';
 import {
+  SimulationErrorCode,
   SimulationTokenStandard,
   TransactionStatus,
   TransactionType,
@@ -61,7 +62,10 @@ import {
 } from './types';
 import { addGasBuffer, estimateGas, updateGas } from './utils/gas';
 import { updateGasFees } from './utils/gas-fees';
-import { updateTransactionLayer1GasFee } from './utils/layer1-gas-fee-flow';
+import {
+  getTransactionLayer1GasFee,
+  updateTransactionLayer1GasFee,
+} from './utils/layer1-gas-fee-flow';
 import { getSimulationData } from './utils/simulation';
 import {
   updatePostTransactionBalance,
@@ -749,6 +753,9 @@ describe('TransactionController', () => {
         getEthQuery: jest.fn().mockImplementation(() => {
           return new EthQuery(provider);
         }),
+        getProvider: jest.fn().mockImplementation(() => {
+          return provider;
+        }),
         checkForPendingTransactionAndStartPolling: jest.fn(),
         getNonceLock: getNonceLockSpy,
         initialize: jest.fn(),
@@ -808,6 +815,18 @@ describe('TransactionController', () => {
           gasFeeFlows: [lineaGasFeeFlowMock],
         }),
       );
+    });
+
+    it('checks pending transactions', () => {
+      expect(
+        pendingTransactionTrackerMock.startIfPendingTransactions,
+      ).toHaveBeenCalledTimes(0);
+
+      setupController();
+
+      expect(
+        pendingTransactionTrackerMock.startIfPendingTransactions,
+      ).toHaveBeenCalledTimes(1);
     });
 
     describe('nonce tracker', () => {
@@ -1378,7 +1397,6 @@ describe('TransactionController', () => {
         origin: undefined,
         securityAlertResponse: undefined,
         sendFlowHistory: expect.any(Array),
-        simulationData: undefined,
         status: TransactionStatus.unapproved as const,
         time: expect.any(Number),
         txParams: expect.anything(),
@@ -1686,44 +1704,71 @@ describe('TransactionController', () => {
       });
     });
 
-    it('updates simulation data by default', async () => {
-      getSimulationDataMock.mockResolvedValueOnce(SIMULATION_DATA_MOCK);
+    describe('updates simulation data', () => {
+      it('by default', async () => {
+        getSimulationDataMock.mockResolvedValueOnce(SIMULATION_DATA_MOCK);
 
-      const { controller } = setupController();
+        const { controller } = setupController();
 
-      await controller.addTransaction({
-        from: ACCOUNT_MOCK,
-        to: ACCOUNT_MOCK,
+        await controller.addTransaction({
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        });
+
+        await flushPromises();
+
+        expect(getSimulationDataMock).toHaveBeenCalledTimes(1);
+        expect(getSimulationDataMock).toHaveBeenCalledWith({
+          chainId: MOCK_NETWORK.state.providerConfig.chainId,
+          data: undefined,
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+          value: '0x0',
+        });
+
+        expect(controller.state.transactions[0].simulationData).toStrictEqual(
+          SIMULATION_DATA_MOCK,
+        );
       });
 
-      expect(getSimulationDataMock).toHaveBeenCalledTimes(1);
-      expect(getSimulationDataMock).toHaveBeenCalledWith({
-        chainId: MOCK_NETWORK.state.providerConfig.chainId,
-        data: undefined,
-        from: ACCOUNT_MOCK,
-        to: ACCOUNT_MOCK,
-        value: '0x0',
+      it('with error if simulation disabled', async () => {
+        getSimulationDataMock.mockResolvedValueOnce(SIMULATION_DATA_MOCK);
+
+        const { controller } = setupController({
+          options: { isSimulationEnabled: () => false },
+        });
+
+        await controller.addTransaction({
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        });
+
+        expect(getSimulationDataMock).toHaveBeenCalledTimes(0);
+        expect(controller.state.transactions[0].simulationData).toStrictEqual({
+          error: {
+            code: SimulationErrorCode.Disabled,
+            message: 'Simulation disabled',
+          },
+          tokenBalanceChanges: [],
+        });
       });
 
-      expect(controller.state.transactions[0].simulationData).toStrictEqual(
-        SIMULATION_DATA_MOCK,
-      );
-    });
+      it('unless approval not required', async () => {
+        getSimulationDataMock.mockResolvedValueOnce(SIMULATION_DATA_MOCK);
 
-    it('does not update simulation data if simulation disabled', async () => {
-      getSimulationDataMock.mockResolvedValueOnce(SIMULATION_DATA_MOCK);
+        const { controller } = setupController();
 
-      const { controller } = setupController({
-        options: { isSimulationEnabled: () => false },
+        await controller.addTransaction(
+          {
+            from: ACCOUNT_MOCK,
+            to: ACCOUNT_MOCK,
+          },
+          { requireApproval: false },
+        );
+
+        expect(getSimulationDataMock).toHaveBeenCalledTimes(0);
+        expect(controller.state.transactions[0].simulationData).toBeUndefined();
       });
-
-      await controller.addTransaction({
-        from: ACCOUNT_MOCK,
-        to: ACCOUNT_MOCK,
-      });
-
-      expect(getSimulationDataMock).toHaveBeenCalledTimes(0);
-      expect(controller.state.transactions[0].simulationData).toBeUndefined();
     });
 
     describe('on approve', () => {
@@ -4404,7 +4449,7 @@ describe('TransactionController', () => {
 
       expect(signSpy).toHaveBeenCalledTimes(1);
 
-      expect(updateTransactionSpy).toHaveBeenCalledTimes(2);
+      expect(updateTransactionSpy).toHaveBeenCalledTimes(3);
       expect(updateTransactionSpy).toHaveBeenCalledWith(
         expect.objectContaining({
           txParams: expect.objectContaining(paramsMock),
@@ -4448,7 +4493,7 @@ describe('TransactionController', () => {
       await wait(0);
 
       expect(signSpy).toHaveBeenCalledTimes(1);
-      expect(updateTransactionSpy).toHaveBeenCalledTimes(1);
+      expect(updateTransactionSpy).toHaveBeenCalledTimes(2);
     });
 
     it('adds a transaction, signs and skips publish the transaction', async () => {
@@ -4474,7 +4519,7 @@ describe('TransactionController', () => {
 
       expect(signSpy).toHaveBeenCalledTimes(1);
 
-      expect(updateTransactionSpy).toHaveBeenCalledTimes(2);
+      expect(updateTransactionSpy).toHaveBeenCalledTimes(3);
       expect(updateTransactionSpy).toHaveBeenCalledWith(
         expect.objectContaining({
           txParams: expect.objectContaining(paramsMock),
@@ -4544,6 +4589,37 @@ describe('TransactionController', () => {
       );
 
       expect(mockEthQuery.sendRawTransaction).toHaveBeenCalledTimes(1);
+    });
+
+    it('submits to publish hook with final transaction meta', async () => {
+      const publishHook = jest
+        .fn()
+        .mockResolvedValue({ transactionHash: TRANSACTION_META_MOCK.hash });
+
+      const { controller } = setupController({
+        options: {
+          hooks: {
+            publish: publishHook,
+          },
+        },
+        messengerOptions: {
+          addTransactionApprovalRequest: {
+            state: 'approved',
+          },
+        },
+      });
+
+      const { result } = await controller.addTransaction(paramsMock);
+
+      await result;
+
+      expect(publishHook).toHaveBeenCalledTimes(1);
+      expect(publishHook).toHaveBeenCalledWith(
+        expect.objectContaining({
+          txParams: expect.objectContaining({ nonce: toHex(NONCE_MOCK) }),
+        }),
+        expect.any(String),
+      );
     });
   });
 
@@ -5487,6 +5563,34 @@ describe('TransactionController', () => {
         .toThrow(`TransactionsController: Can only call updateEditableParams on an unapproved transaction.
       Current tx status: ${TransactionStatus.submitted}`);
     });
+
+    it.each(['value', 'to', 'data'])(
+      'updates simulation data if %s changes',
+      async (param) => {
+        const { controller } = setupController({
+          options: {
+            state: {
+              transactions: [
+                {
+                  ...transactionMeta,
+                },
+              ],
+            },
+          },
+        });
+
+        expect(getSimulationDataMock).toHaveBeenCalledTimes(0);
+
+        await controller.updateEditableParams(transactionMeta.id, {
+          ...transactionMeta.txParams,
+          [param]: ACCOUNT_2_MOCK,
+        });
+
+        await flushPromises();
+
+        expect(getSimulationDataMock).toHaveBeenCalledTimes(1);
+      },
+    );
   });
 
   describe('abortTransactionSigning', () => {
@@ -5551,6 +5655,35 @@ describe('TransactionController', () => {
           }
         ).error.message,
       ).toBe('Signing aborted by user');
+    });
+  });
+
+  describe('getLayer1GasFee', () => {
+    it('calls getLayer1GasFee with the correct parameters', async () => {
+      const chainIdMock = '0x1';
+      const networkClientIdMock = 'mainnet';
+      const transactionParamsMock = {
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        gas: '0x0',
+        gasPrice: '0x50fd51da',
+        value: '0x0',
+      };
+      const layer1GasFeeMock = '0x12356';
+      (getTransactionLayer1GasFee as jest.Mock).mockResolvedValueOnce(
+        layer1GasFeeMock,
+      );
+
+      const { controller } = setupController();
+
+      expect(
+        await controller.getLayer1GasFee(
+          chainIdMock,
+          networkClientIdMock,
+          transactionParamsMock,
+        ),
+      ).toBe(layer1GasFeeMock);
+      expect(getTransactionLayer1GasFee).toHaveBeenCalledTimes(1);
     });
   });
 });