+++ packages/transaction-controller/src/TransactionController.ts
@@ -39,7 +39,7 @@ import { add0x } from '@metamask/utils';
 import { Mutex } from 'async-mutex';
 import { MethodRegistry } from 'eth-method-registry';
 import { EventEmitter } from 'events';
-import { cloneDeep, mapValues, merge, pickBy, sortBy } from 'lodash';
+import { cloneDeep, mapValues, merge, pickBy, sortBy, isEqual } from 'lodash';
 import { NonceTracker } from 'nonce-tracker';
 import type {
   NonceLock,
@@ -49,6 +49,7 @@ import { v1 as random } from 'uuid';
 
 import { DefaultGasFeeFlow } from './gas-flows/DefaultGasFeeFlow';
 import { LineaGasFeeFlow } from './gas-flows/LineaGasFeeFlow';
+import { OptimismLayer1GasFeeFlow } from './gas-flows/OptimismLayer1GasFeeFlow';
 import { EtherscanRemoteTransactionSource } from './helpers/EtherscanRemoteTransactionSource';
 import { GasFeePoller } from './helpers/GasFeePoller';
 import type { IncomingTransactionOptions } from './helpers/IncomingTransactionHelper';
@@ -68,11 +69,13 @@ import type {
   WalletDevice,
   SecurityAlertResponse,
   GasFeeFlow,
+  SimulationData,
 } from './types';
 import {
   TransactionEnvelopeType,
   TransactionType,
   TransactionStatus,
+  SimulationErrorCode,
 } from './types';
 import { validateConfirmedExternalTransaction } from './utils/external-transactions';
 import { addGasBuffer, estimateGas, updateGas } from './utils/gas';
@@ -81,7 +84,10 @@ import {
   addInitialHistorySnapshot,
   updateTransactionHistory,
 } from './utils/history';
-import { updateTransactionLayer1GasFee } from './utils/layer1-gas-fee-flow';
+import {
+  getTransactionLayer1GasFee,
+  updateTransactionLayer1GasFee,
+} from './utils/layer1-gas-fee-flow';
 import {
   getAndFormatTransactionsForNonceTracker,
   getNextNonce,
@@ -227,7 +233,7 @@ export type TransactionControllerActions = TransactionControllerGetStateAction;
  * @property isResubmitEnabled - Whether transaction publishing is automatically retried.
  */
 export type PendingTransactionOptions = {
-  isResubmitEnabled?: boolean;
+  isResubmitEnabled?: () => boolean;
 };
 
 /**
@@ -853,12 +859,12 @@ export class TransactionController extends BaseController<
     const gasFeePoller = new GasFeePoller({
       // Default gas fee polling is not yet supported by the clients
       gasFeeFlows: this.gasFeeFlows.slice(0, -1),
-      getEthQuery: (chainId, networkClientId) =>
-        this.#multichainTrackingHelper.getEthQuery({
+      getGasFeeControllerEstimates: this.getGasFeeEstimates,
+      getProvider: (chainId, networkClientId) =>
+        this.#multichainTrackingHelper.getProvider({
           networkClientId,
           chainId,
         }),
-      getGasFeeControllerEstimates: this.getGasFeeEstimates,
       getTransactions: () => this.state.transactions,
       layer1GasFeeFlows: this.layer1GasFeeFlows,
       onStateChange: (listener) => {
@@ -889,6 +895,7 @@ export class TransactionController extends BaseController<
     });
 
     this.onBootCleanup();
+    this.#checkForPendingTransactionAndStartPolling();
   }
 
   /**
@@ -1038,10 +1045,7 @@ export class TransactionController extends BaseController<
           networkClientId,
         };
 
-    await Promise.all([
-      this.updateGasProperties(addedTransactionMeta),
-      this.#simulateTransaction(addedTransactionMeta),
-    ]);
+    await this.updateGasProperties(addedTransactionMeta);
 
     // Checks if a transaction already exists with a given actionId
     if (!existingTransactionMeta) {
@@ -1075,6 +1079,14 @@ export class TransactionController extends BaseController<
       );
 
       this.addMetadata(addedTransactionMeta);
+
+      if (requireApproval !== false) {
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
+        this.#updateSimulationData(addedTransactionMeta);
+      } else {
+        log('Skipping simulation as approval not required');
+      }
+
       this.messagingSystem.publish(
         `${controllerName}:unapprovedTransactionAdded`,
         addedTransactionMeta,
@@ -1929,10 +1941,11 @@ export class TransactionController extends BaseController<
     ) as TransactionParams;
 
     const updatedTransaction = merge({}, transactionMeta, editableParams);
-    const ethQuery = this.#multichainTrackingHelper.getEthQuery({
-      networkClientId: transactionMeta.networkClientId,
+    const provider = this.#multichainTrackingHelper.getProvider({
       chainId: transactionMeta.chainId,
+      networkClientId: transactionMeta.networkClientId,
     });
+    const ethQuery = new EthQuery(provider);
     const { type } = await determineTransactionType(
       updatedTransaction.txParams,
       ethQuery,
@@ -1940,8 +1953,8 @@ export class TransactionController extends BaseController<
     updatedTransaction.type = type;
 
     await updateTransactionLayer1GasFee({
-      ethQuery,
       layer1GasFeeFlows: this.layer1GasFeeFlows,
+      provider,
       transactionMeta: updatedTransaction,
     });
 
@@ -2111,6 +2124,10 @@ export class TransactionController extends BaseController<
         `${controllerName}:transactionFinished`,
         updatedTransactionMeta,
       );
+      this.#internalEvents.emit(
+        `${updatedTransactionMeta.id}:finished`,
+        updatedTransactionMeta,
+      );
     }
   }
 
@@ -2241,6 +2258,34 @@ export class TransactionController extends BaseController<
     return filteredTransactions;
   }
 
+  /**
+   * Utility method to get the layer 1 gas fee for given transaction params.
+   *
+   * @param chainId - Estimated transaction chainId.
+   * @param networkClientId - Estimated transaction networkClientId.
+   * @param transactionParams - The transaction params to estimate layer 1 gas fee for.
+   */
+  async getLayer1GasFee(
+    chainId: Hex,
+    networkClientId: NetworkClientId,
+    transactionParams: TransactionParams,
+  ): Promise<Hex | undefined> {
+    const provider = this.#multichainTrackingHelper.getProvider({
+      networkClientId,
+      chainId,
+    });
+
+    const layer1GasFee = await getTransactionLayer1GasFee({
+      layer1GasFeeFlows: this.layer1GasFeeFlows,
+      provider,
+      transactionMeta: {
+        txParams: transactionParams,
+        chainId,
+      } as TransactionMeta,
+    });
+    return layer1GasFee;
+  }
+
   private async signExternalTransaction(
     chainId: Hex,
     transactionParams: TransactionParams,
@@ -2338,6 +2383,11 @@ export class TransactionController extends BaseController<
       chainId,
     });
 
+    const provider = this.#multichainTrackingHelper.getProvider({
+      networkClientId,
+      chainId,
+    });
+
     await updateGas({
       ethQuery,
       chainId,
@@ -2355,8 +2405,8 @@ export class TransactionController extends BaseController<
     });
 
     await updateTransactionLayer1GasFee({
-      ethQuery,
       layer1GasFeeFlows: this.layer1GasFeeFlows,
+      provider,
       transactionMeta,
     });
   }
@@ -2614,7 +2664,7 @@ export class TransactionController extends BaseController<
       log('Publishing transaction', txParams);
 
       let { transactionHash: hash } = await this.publish(
-        transactionMeta,
+        updatedTransactionMeta,
         rawTx,
       );
 
@@ -3478,7 +3528,7 @@ export class TransactionController extends BaseController<
   }
 
   #getLayer1GasFeeFlows(): Layer1GasFeeFlow[] {
-    return [];
+    return [new OptimismLayer1GasFeeFlow()];
   }
 
   #updateTransactionInternal(
@@ -3491,6 +3541,10 @@ export class TransactionController extends BaseController<
     };
 
     validateTxParams(normalizedTransaction.txParams);
+
+    const updatedTransactionParams = this.#checkIfTransactionParamsUpdated(
+      normalizedTransaction,
+    );
 
     const transactionWithUpdatedHistory =
       skipHistory === true
@@ -3506,25 +3560,100 @@ export class TransactionController extends BaseController<
       );
       state.transactions[index] = transactionWithUpdatedHistory;
     });
+
+    if (updatedTransactionParams.length > 0) {
+      this.#onTransactionParamsUpdated(
+        normalizedTransaction,
+        updatedTransactionParams,
+      );
+    }
   }
 
-  async #simulateTransaction(transactionMeta: TransactionMeta) {
-    if (!this.#isSimulationEnabled()) {
-      log('Skipping simulation as disabled');
+  #checkIfTransactionParamsUpdated(newTransactionMeta: TransactionMeta) {
+    const { id: transactionId, txParams: newParams } = newTransactionMeta;
+
+    const originalParams = this.getTransaction(transactionId)?.txParams;
+
+    if (!originalParams || isEqual(originalParams, newParams)) {
+      return [];
+    }
+
+    const params = Object.keys(newParams) as (keyof TransactionParams)[];
+
+    const updatedProperties = params.filter(
+      (param) => newParams[param] !== originalParams[param],
+    );
+
+    log(
+      'Transaction parameters have been updated',
+      transactionId,
+      updatedProperties,
+      originalParams,
+      newParams,
+    );
+
+    return updatedProperties;
+  }
+
+  #onTransactionParamsUpdated(
+    transactionMeta: TransactionMeta,
+    updatedParams: (keyof TransactionParams)[],
+  ) {
+    if (
+      (['to', 'value', 'data'] as const).some((param) =>
+        updatedParams.includes(param),
+      )
+    ) {
+      log('Updating simulation data due to transaction parameter update');
+      // eslint-disable-next-line @typescript-eslint/no-floating-promises
+      this.#updateSimulationData(transactionMeta);
+    }
+  }
+
+  async #updateSimulationData(transactionMeta: TransactionMeta) {
+    const { id, chainId, txParams } = transactionMeta;
+    const { from, to, value, data } = txParams;
+
+    let simulationData: SimulationData = {
+      error: {
+        code: SimulationErrorCode.Disabled,
+        message: 'Simulation disabled',
+      },
+      tokenBalanceChanges: [],
+    };
+
+    if (this.#isSimulationEnabled()) {
+      this.#updateTransactionInternal(
+        { ...transactionMeta, simulationData: undefined },
+        { skipHistory: true },
+      );
+
+      simulationData = await getSimulationData({
+        chainId,
+        from: from as Hex,
+        to: to as Hex,
+        value: value as Hex,
+        data: data as Hex,
+      });
+    }
+
+    const finalTransactionMeta = this.getTransaction(id);
+
+    if (!finalTransactionMeta) {
+      log(
+        'Cannot update simulation data as transaction not found',
+        id,
+        simulationData,
+      );
+
       return;
     }
 
-    const { chainId, txParams } = transactionMeta;
-    const { from, to, value, data } = txParams;
+    this.updateTransaction(
+      { ...finalTransactionMeta, simulationData },
+      'TransactionController#updateSimulationData - Update simulation data',
+    );
 
-    transactionMeta.simulationData = await getSimulationData({
-      chainId,
-      from: from as Hex,
-      to: to as Hex,
-      value: value as Hex,
-      data: data as Hex,
-    });
-
-    log('Retrieved simulation data', transactionMeta.simulationData);
+    log('Updated simulation data', id, simulationData);
   }
 }