+++ packages/transaction-controller/src/utils/simulation-api.ts
@@ -1,40 +1,37 @@
-import { JsonRpcProvider } from '@ethersproject/providers';
+import { convertHexToDecimal } from '@metamask/controller-utils';
 import { createModuleLogger, type Hex } from '@metamask/utils';
 
-import { CHAIN_IDS } from '../constants';
+import { SimulationChainNotSupportedError, SimulationError } from '../errors';
 import { projectLogger } from '../logger';
 
 const log = createModuleLogger(projectLogger, 'simulation-api');
 
 const RPC_METHOD = 'infura_simulateTransactions';
 const BASE_URL = 'https://tx-sentinel-{0}.api.cx.metamask.io/';
-
-const SUBDOMAIN_BY_CHAIN_ID: Record<Hex, string> = {
-  [CHAIN_IDS.MAINNET]: 'ethereum-mainnet',
-  [CHAIN_IDS.GOERLI]: 'ethereum-goerli',
-  [CHAIN_IDS.SEPOLIA]: 'ethereum-sepolia',
-  [CHAIN_IDS.LINEA_MAINNET]: 'linea-mainnet',
-  [CHAIN_IDS.LINEA_GOERLI]: 'linea-goerli',
-  [CHAIN_IDS.ARBITRUM]: 'arbitrum-mainnet',
-  [CHAIN_IDS.AVALANCHE]: 'avalanche-mainnet',
-  [CHAIN_IDS.OPTIMISM]: 'optimism-mainnet',
-  [CHAIN_IDS.POLYGON]: 'polygon-mainnet',
-  [CHAIN_IDS.BSC]: 'bsc-mainnet',
-};
+const ENDPOINT_NETWORKS = 'networks';
 
 /** Single transaction to simulate in a simulation API request.  */
 export type SimulationRequestTransaction = {
+  /** Data to send with the transaction. */
+  data?: Hex;
+
   /** Sender of the transaction. */
   from: Hex;
+
+  /** Gas limit for the transaction. */
+  gas?: Hex;
+
+  /** Maximum fee per gas for the transaction. */
+  maxFeePerGas?: Hex;
+
+  /** Maximum priority fee per gas for the transaction. */
+  maxPriorityFeePerGas?: Hex;
 
   /** Recipient of the transaction. */
   to?: Hex;
 
   /** Value to send with the transaction. */
   value?: Hex;
-
-  /** Data to send with the transaction. */
-  data?: Hex;
 };
 
 /** Request to the simulation API to simulate transactions. */
@@ -112,19 +109,22 @@ export type SimulationResponseStateDiff = {
 
 /** Response from the simulation API for a single transaction. */
 export type SimulationResponseTransaction = {
+  /** An error message indicating the transaction could not be simulated. */
+  error?: string;
+
   /** Return value of the transaction, such as the balance if calling balanceOf. */
   return: Hex;
 
   /** Hierarchy of call data including nested calls and logs. */
-  callTrace: SimulationResponseCallTrace;
+  callTrace?: SimulationResponseCallTrace;
 
   /** Changes to the blockchain state. */
-  stateDiff: {
+  stateDiff?: {
     /** Initial blockchain state before the transaction. */
-    pre: SimulationResponseStateDiff;
+    pre?: SimulationResponseStateDiff;
 
     /** Updated blockchain state after the transaction. */
-    post: SimulationResponseStateDiff;
+    post?: SimulationResponseStateDiff;
   };
 };
 
@@ -133,6 +133,22 @@ export type SimulationResponse = {
   /** Simulation data for each transaction in the request. */
   transactions: SimulationResponseTransaction[];
 };
+
+/** Data for a network supported by the Simulation API. */
+type SimulationNetwork = {
+  /** Subdomain of the API for the network.  */
+  network: string;
+
+  /** Whether the network supports confirmation simulations. */
+  confirmations: boolean;
+};
+
+/** Response from the simulation API containing supported networks. */
+type SimulationNetworkResponse = {
+  [chainIdDecimal: string]: SimulationNetwork;
+};
+
+let requestIdCounter = 0;
 
 /**
  * Simulate transactions using the transaction simulation API.
@@ -143,16 +159,33 @@ export async function simulateTransactions(
   chainId: Hex,
   request: SimulationRequest,
 ): Promise<SimulationResponse> {
-  const url = getUrl(chainId);
+  const url = await getSimulationUrl(chainId);
 
   log('Sending request', url, request);
 
-  const jsonRpc = new JsonRpcProvider(url);
-  const response = await jsonRpc.send(RPC_METHOD, [request]);
+  const requestId = requestIdCounter;
+  requestIdCounter += 1;
 
-  log('Received response', response);
+  const response = await fetch(url, {
+    method: 'POST',
+    body: JSON.stringify({
+      id: String(requestId),
+      jsonrpc: '2.0',
+      method: RPC_METHOD,
+      params: [request],
+    }),
+  });
 
-  return response;
+  const responseJson = await response.json();
+
+  log('Received response', responseJson);
+
+  if (responseJson.error) {
+    const { code, message } = responseJson.error;
+    throw new SimulationError(message, code);
+  }
+
+  return responseJson?.result;
 }
 
 /**
@@ -160,13 +193,33 @@ export async function simulateTransactions(
  * @param chainId - The chain ID to get the URL for.
  * @returns The URL for the transaction simulation API.
  */
-function getUrl(chainId: Hex): string {
-  const subdomain = SUBDOMAIN_BY_CHAIN_ID[chainId];
+async function getSimulationUrl(chainId: Hex): Promise<string> {
+  const networkData = await getNetworkData();
+  const chainIdDecimal = convertHexToDecimal(chainId);
+  const network = networkData[chainIdDecimal];
 
-  if (!subdomain) {
+  if (!network?.confirmations) {
     log('Chain is not supported', chainId);
-    throw new Error(`Chain is not supported: ${chainId}`);
+    throw new SimulationChainNotSupportedError(chainId);
   }
 
+  return getUrl(network.network);
+}
+
+/**
+ * Retrieve the supported network data from the simulation API.
+ */
+async function getNetworkData(): Promise<SimulationNetworkResponse> {
+  const url = `${getUrl('ethereum-mainnet')}${ENDPOINT_NETWORKS}`;
+  const response = await fetch(url);
+  return response.json();
+}
+
+/**
+ * Generate the URL for the specified subdomain in the simulation API.
+ * @param subdomain - The subdomain to generate the URL for.
+ * @returns The URL for the transaction simulation API.
+ */
+function getUrl(subdomain: string): string {
   return BASE_URL.replace('{0}', subdomain);
 }